Compute Primitives together with full resolution spline samples
This solves the problem of the wrong index computation,
due to rounding errors, while maintaining full resolution
splines for the final trajectory.

Also the position of the collision check is now computed as
near to the cell middle as possible.

The intermediate steps now contain the full path, with the final
position.