Distribute goal position offset along the discrete positions of the final trajectory
The planner did not always give a trajectory to the exact goal position. This was
due to the discrete resolution uses to generate the primitives. For example: if the
primitives are generated with a grid resolution of 0.3 then the planner final trajectory
position has an offset of 0.3 or less for the actual goal position